+++ app/src/main/java/com/sdu/composemusicplayer/MainActivity.kt
@@ -12,7 +12,6 @@ import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.tooling.preview.PreviewScreenSizes
 import com.sdu.composemusicplayer.navigation.SetupNavigation
 import com.sdu.composemusicplayer.presentation.permission.CheckAndRequestPermissions
