+++ app/src/main/java/com/sdu/composemusicplayer/presentation/musicPlayerSheet/compoenent/AlbumImage.kt
@@ -1,4 +1,4 @@
-package com.sdu.composemusicplayer.presentation.music_player_sheet.compoenent
+package com.sdu.composemusicplayer.presentation.musicPlayerSheet.compoenent
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.fillMaxSize
