+++ app/src/main/java/com/sdu/composemusicplayer/presentation/musicPlayerSheet/compoenent/MotionContent.kt
@@ -1,6 +1,6 @@
 @file:OptIn(ExperimentalMotionApi::class)
 
-package com.sdu.composemusicplayer.presentation.music_player_sheet.compoenent
+package com.sdu.composemusicplayer.presentation.musicPlayerSheet.compoenent
 
 import android.util.Log
 import androidx.compose.animation.AnimatedVisibility
