+++ app/src/main/java/com/sdu/composemusicplayer/utils/AppStateUtil.kt
@@ -1,6 +1,5 @@
 package com.sdu.composemusicplayer.utils
 
-import android.annotation.SuppressLint
 import android.app.ActivityManager
 import android.content.Context
 
@@ -10,9 +9,6 @@ object AppStateUtil {
         val appProcesses = activityManager.runningAppProcesses ?: return false
         val packageName = context.packageName
 
-        return appProcesses.any { appProcess -> 
-            appProcess.importance == ActivityManager.RunningAppProcessInfo.IMPORTANCE_FOREGROUND 
-            && appProcess.processName == packageName 
-        }
+        return appProcesses.any { appProcess -> appProcess.importance == ActivityManager.RunningAppProcessInfo.IMPORTANCE_FOREGROUND && appProcess.processName == packageName }
     }
 }
