+++ build.gradle.kts
@@ -7,7 +7,6 @@ plugins {
     id("org.jlleitschuh.gradle.ktlint") version "11.3.1"
     id("io.gitlab.arturbosch.detekt") version "1.22.0"
     alias(libs.plugins.android.library) apply false
-
 }
 
 subprojects {
